requirements-selenium.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
pytest-selenium>=4.1.0
pytest-html>=3.2.0
allure-pytest>=2.13.0
